requirements.txt
google-genai
flask
pillow
python-dotenv
authlib
requests
